requirements.txt
python-docx
pydantic
requests
